+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## v0.1.3
+- Add url string encoding to utils
+- Add utils with net address lookup
+- Optimize code and text
+
 ## v0.1.2
 - Add common helpers
 - Initial web api example ipinfo